+++ backend/platform/src/main/java/ru/abondin/hreasy/platform/repo/employee/admin/UserSecurityInfoEntry.java
@@ -14,13 +14,25 @@ public class UserSecurityInfoEntry extends EmployeeEntry {
      */
     private List<Integer> accessibleProjects;
     /**
+     * List of projects where employee is assigned as manager
+     */
+    private List<Integer> managedProjects;
+    /**
      * List of accessible departments
      */
     private List<Integer> accessibleDepartments;
     /**
+     * List of departments where employee is assigned as manager
+     */
+    private List<Integer> managedDepartments;
+    /**
      * List of accessible business accounts
      */
     private List<Integer> accessibleBas;
+    /**
+     * List of business accounts where employee is assigned as manager
+     */
+    private List<Integer> managedBas;
     /**
      * List of roles
      */