+++ web/src/services/admin/admin-user.service.ts
@@ -10,6 +10,9 @@ export interface UserSecurityInfo {
   accessibleDepartments: number[];
   accessibleProjects: number[];
   accessibleBas: number[];
+  managedDepartments: number[];
+  managedProjects: number[];
+  managedBas: number[];
   dateOfDismissal: string;
 }